abstract log-codes as module over 500 limit
