+++ src/server/auth/services/whitelist.service.js
@@ -122,7 +122,7 @@ class WhitelistService {
   /**
    * Helper method to log whitelist events with consistent structure
    * @param {object} logCode - Logging options.
-   * @param {import('../../common/helpers/logging/log-codes-definition.js').LogTypes.LogLevel} logCode.level - The log level.
+   * @param {import('../../common/helpers/logging/log-codes/definition.js').LogTypes.LogLevel} logCode.level - The log level.
    * @param {Function} logCode.messageFunc - A function that creates an interpolated message string
    * @param {string} crn - The user's CRN
    * @param {string} sbi - The SBI number
