+++ src/server/common/helpers/logging/log-codes/cookies.js
@@ -1,0 +1,10 @@
+/**
+ * @type {Object<string, import('./definition.js').LogCodesDefinition>}
+ */
+export const COOKIES = {
+  PAGE_LOAD: {
+    level: 'info',
+    messageFunc: (messageOptions) =>
+      `Cookies page loaded: returnUrl=${messageOptions.returnUrl}, referer=${messageOptions.referer}`
+  }
+}
