+++ src/server/common/helpers/logging/log-codes/woodland.js
@@ -1,0 +1,9 @@
+/**
+ * @type {Object<string, import('./definition.js').LogCodesDefinition>}
+ */
+export const WOODLAND = {
+  VALIDATE_ERROR: {
+    level: 'error',
+    messageFunc: (messageOptions) => `Woodland validation error: ${messageOptions.errorMessage}`
+  }
+}
