+++ src/server/common/helpers/logging/log.js
@@ -6,7 +6,7 @@ const logger = pino(loggerOptions)
 /**
  * Logs an event with the specified level and context.
  * @param {object} logCode - Logging options.
- * @param {import('./log-codes-definition.js').LogTypes.LogLevel} logCode.level - The log level.
+ * @param {import('./log-codes/definition.js').LogTypes.LogLevel} logCode.level - The log level.
  * @param {Function} logCode.messageFunc - A function that creates an interpolated message string
  * @param {object} [logCode.error] - An error object (optional)
  * @param {object} messageOptions - Values for message interpolation
@@ -21,7 +21,7 @@ const log = (logCode, messageOptions, request) => {
 /**
  * Logs an event with the specified level and context.
  * @param {object} logCode - Logging options.
- * @param {import('./log-codes-definition.js').LogTypes.LogLevel} logCode.level - The log level.
+ * @param {import('./log-codes/definition.js').LogTypes.LogLevel} logCode.level - The log level.
  * @param {Function} logCode.messageFunc - A function that creates an interpolated message string
  * @param {object} [logCode.error] - An error object (optional)
  * @param {object} messageOptions - Values for message interpolation
@@ -51,7 +51,7 @@ const error = (logCode, messageOptions, request) => {
 
 /**
  * Returns the logger function corresponding to the given log level.
- * @param {import('./log-codes-definition.js').LogTypes.LogLevel} level - The log level.
+ * @param {import('./log-codes/definition.js').LogTypes.LogLevel} level - The log level.
  * @param {object} [request] - Hapi request object (optional)
  * @returns {(errorContext: object | undefined, message: string) => void} Logger function.
  */
